refactor: modularize e2e test helpers into separate modules
Split the monolithic test_helper.ts into focused modules:
- constants.ts: Timeout constants
- fixtures.ts: Playwright fixtures and test setup
- utils/: Utility functions (api-request-helpers, electron-helpers, etc.)
- page-objects/: Page object model implementations

This improves maintainability and makes the test helpers easier to navigate.
