+++ e2e-tests/helpers/page-objects/components/Navigation.ts
@@ -1,0 +1,45 @@
+/**
+ * Page object for navigation between tabs and pages.
+ * Handles tab navigation and back button.
+ */
+
+import { Page, expect } from "@playwright/test";
+
+export class Navigation {
+  constructor(public page: Page) {}
+
+  async goToSettingsTab() {
+    await this.page.getByRole("link", { name: "Settings" }).click();
+  }
+
+  async goToLibraryTab() {
+    await this.page.getByRole("link", { name: "Library" }).click();
+  }
+
+  async goToAppsTab() {
+    await this.page.getByRole("link", { name: "Apps" }).click();
+    await expect(this.page.getByText("Build a new app")).toBeVisible();
+  }
+
+  async goToChatTab() {
+    await this.page.getByRole("link", { name: "Chat" }).click();
+  }
+
+  async goToHubTab() {
+    await this.page.getByRole("link", { name: "Hub" }).click();
+  }
+
+  async clickBackButton() {
+    await this.page.getByRole("button", { name: "Back" }).click();
+  }
+
+  async selectTemplate(templateName: string) {
+    await this.page.getByRole("img", { name: templateName }).click();
+  }
+
+  async goToHubAndSelectTemplate(templateName: "Next.js Template") {
+    await this.goToHubTab();
+    await this.selectTemplate(templateName);
+    await this.goToAppsTab();
+  }
+}
