+++ e2e-tests/helpers/page-objects/components/index.ts
@@ -1,0 +1,16 @@
+/**
+ * Barrel file for component page objects.
+ */
+
+export { GitHubConnector } from "./GitHubConnector";
+export { ChatActions } from "./ChatActions";
+export { PreviewPanel } from "./PreviewPanel";
+export { CodeEditor } from "./CodeEditor";
+export { SecurityReview } from "./SecurityReview";
+export { ToastNotifications } from "./ToastNotifications";
+export { AgentConsent } from "./AgentConsent";
+export { Navigation } from "./Navigation";
+export { ModelPicker } from "./ModelPicker";
+export { Settings } from "./Settings";
+export { AppManagement } from "./AppManagement";
+export { PromptLibrary } from "./PromptLibrary";
