+++ e2e-tests/helpers/page-objects/dialogs/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Barrel file for dialog page objects.
+ */
+
+export { ContextFilesPickerDialog } from "./ContextFilesPickerDialog";
+export { ProModesDialog } from "./ProModesDialog";
