+++ e2e-tests/helpers/page-objects/index.ts
@@ -1,0 +1,25 @@
+/**
+ * Barrel file for page objects.
+ */
+
+// Main page object
+export { PageObject } from "./PageObject";
+
+// Dialog page objects
+export { ContextFilesPickerDialog, ProModesDialog } from "./dialogs";
+
+// Component page objects
+export {
+  GitHubConnector,
+  ChatActions,
+  PreviewPanel,
+  CodeEditor,
+  SecurityReview,
+  ToastNotifications,
+  AgentConsent,
+  Navigation,
+  ModelPicker,
+  Settings,
+  AppManagement,
+  PromptLibrary,
+} from "./components";
