+++ e2e-tests/helpers/utils/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Barrel file for utility exports.
+ */
+
+export {
+  normalizeItemReferences,
+  normalizeToolCallIds,
+  normalizeVersionedFiles,
+  normalizePath,
+} from "./normalization";
+
+export { prettifyDump, type PrettifyDumpOptions } from "./dump-prettifier";
